added Extras section with requested details (#3)
## Description
<!-- A short one or two line description (like a TLDR) of these changes
-->

Added requested additions and a few missing screen shots from before. :)

## Changes
<!-- An itemized list of changes in this PR

For each change, start with the word MAJOR, MINOR, FIX or CHORE followed
by a colon and brief description of the change.

For example,

* FIX: Correct a big bad bug
* MAJOR: A breaking change that will impact code other devs will write
in the future
* MINOR: Not a big deal, and not likely to affect the work of others,
but something you want shared
* CHORE: Not really a code change, but something done (ie spelling, text
or image updates)
-->

- Added extras section 
- Added screenshots

## To Do
<!-- Use this to track the work you've done, or need to do in this PR
-->


## Outstanding Issues
<!-- If this is a draft PR, there might be items you're looking for
feedback on or need to still address. List them here, and ask for help
if you need it! -->


## See Also
<!-- Does this close or impact an issue. Link to all of them here. -->

Resolves #1 Resolves #2